requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-crispy-forms==1.14.0
django-embed-video==1.4.4
djlint==1.0.1
idna==3.3
importlib-metadata==4.11.3
pathspec==0.9.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
zipp==3.8.0
